p2v: Switch from matchbox-window-manager to metacity.
Matchbox only supports a single window on screen.  This worked fine
until we added the NetworkManager connection dialog -- which tries to
open in a new top level window, and of course fails.

Metacity supports multiple top-level windows so doesn't suffer this
problem.

Apart from that it's hard to tell the difference between the two.