daemon: add split_lines_sb
Mold split_lines_sb from split_lines, so it returns the strings buffer
with the result of the split. This way, we can have the number of lines
in the array, with no need to count them again later.

split_lines is rewritten to take the ownership of the result of
split_lines_sb.