v2v: Catch real exception thrown by failing aug_get (RHBZ#1239053).
When converting the old Perl virt-v2v code, I made a silly mistake
with the exception that aug_get throws when it doesn't find any node.
It throws a 'Guestfs.Error' exception, not 'Not_found'.

As a result of this, the exception was escaping and the proper error
message was not displayed.  With a malformed grub configuration you
would see this error:

  $ virt-v2v -i disk centos-6.img -o null
  [...]
  virt-v2v: error: libguestfs error: aug_get: no matching node

After applying this patch:

  $ virt-v2v -i disk centos-6.img -o null
  [...]
  virt-v2v: error: no kernels were found in the grub configuration.

  This probably indicates that virt-v2v was unable to parse the grub
  configuration of this guest.

which is the correct error message.