dib: handle unsetting functions in environment
If a function name, with its trailing parentheses, is in the
environment , trying to unset it will error out with a message of "not
a valid identifier".  Handle this as a special case, and unset it with
the -f option which can handle the parentheses in the supplied
identifier.

(cherry picked from commit c2bb87e8f26e75e7ce71748785380b76b6e63658)