ocaml: Conditionally remove bindtests.* when doing 'make install'.
Commit f6a790b67d7c8f79532f61d20d11ec8d3a4e003b changed the Makefile
so it only builds the tests when you run 'make check'.

If you didn't run 'make check' then (in particular) bindtests won't be
built, and so when you install it cannot be removed.

The simple fix for this is to make the removal of bindtests.*
conditional.

This fixes commit f6a790b67d7c8f79532f61d20d11ec8d3a4e003b.