p2v: Cancel conversion if user closes the running dialog.
If the user tries to close the running dialog during a running
conversion, act as if the cancel button was pressed (cancelling the
conversion).

If the conversion has finished/been cancelled, then this closes the
dialog and exits virt-p2v, as now.

Cancelling the conversion in this way ensures that qemu-nbd and ssh
subprocesses are cleaned up.