build: installcheck: Ensure libguestfs.so is copied.
If you ran 'make installcheck' when libguestfs-devel was not
installed, then the installcheck would fail in many places with:

  gcc: error: ../../src/.libs/libguestfs.so: No such file or directory

We should stop the installcheck script immediately if the .so file was
not found.

Also this commit ensures that src/.libs/libguestfs.so* not existing
before installcheck is not fatal.

(cherry picked from commit 3fa03a30b3ee766d0c09a79802f326b09f6b66a7)