appliance: Reenable ACPI.
Previously I disabled ACPI because it was not necessary and was slow.
However measuring it again now, I can see no significant difference in
performance, and it is necessary if we are going to use DAX.

It's also useful to keep the appliance kernel state as similar to the
ordinary state, and setting acpi=off is a massive change to the way
that the kernel boots.

Therefore, reenable it.

(cherry picked from commit db1f811b29b055625ae1b03cf0daa9fed0d443ac)