drives: force format=raw for /dev/null dummy drives
When creating the dummy scratch disk for a /dev/null drive, force its
format as raw if it was not specified already.  After all, raw was
already the only format allowed for this kind of drives, so it makes
sense to specify the format explicitly and avoid the autodetection in
qemu.