v2v: Add -o discard option when fstrimming.
This *may* be required by some filesystems in order for fstrim to
work.  I'm not actually sure if this is true, but it's what
virt-sparsify --in-place does, and that utility has been tested a lot
more in regards to trimming.

(cherry picked from commit dd22e376c4ec8fd7c78c78f94c60bb763ddcf08f)